+++ source/tools/acpinames/antables.c
@@ -270,7 +270,7 @@ AeBuildLocalTables (
          * Use the external FADT, but we must update the DSDT/FACS addresses
          * as well as the checksum
          */
-        ExternalFadt->Dsdt = DsdtAddress;
+        ExternalFadt->Dsdt = (UINT32) DsdtAddress;
         ExternalFadt->Facs = ACPI_PTR_TO_PHYSADDR (&LocalFACS);
 
         if (ExternalFadt->Header.Length > ACPI_PTR_DIFF (&ExternalFadt->XDsdt, ExternalFadt))
